requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.0.0=pypi_0
astunparse=1.6.3=pypi_0
blas=2.114=openblas
blas-devel=3.9.0=14_osxarm64_openblas
c-ares=1.18.1=h1a28f6b_0
ca-certificates=2022.4.26=hca03da5_0
cachetools=5.1.0=pypi_0
certifi=2021.10.8=py38hca03da5_2
cffi=1.15.0=py38h22df2f2_1
charset-normalizer=2.0.12=pypi_0
cloudpickle=2.0.0=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.30=pypi_0
fasteners=0.17.3=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=h1192e45_0
future=0.18.2=py38hca03da5_1
gast=0.5.3=pypi_0
giflib=5.2.1=h1a28f6b_0
glfw=2.5.3=pypi_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.42.0=py38h95c9599_0
gym=0.23.1=pypi_0
gym-notices=0.0.6=pypi_0
h5py=3.6.0=py38h7fe8675_0
hdf5=1.12.1=h5aa262f_1
idna=3.3=pypi_0
imageio=2.19.2=pypi_0
importlib-metadata=4.11.3=pypi_0
jpeg=9e=h1a28f6b_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=pypi_0
krb5=1.19.2=h3b8d789_0
lcms2=2.12=hba8e193_0
libblas=3.9.0=14_osxarm64_openblas
libcblas=3.9.0=14_osxarm64_openblas
libclang=14.0.1=pypi_0
libcurl=7.82.0=hc6d1d07_0
libcxx=12.0.0=hf6beb65_1
libedit=3.1.20210910=h1a28f6b_0
libev=4.33=h1a28f6b_1
libffi=3.4.2=hc377ac9_2
libgfortran=5.0.0=11_2_0_he6877d6_26
libgfortran5=11.2.0=he6877d6_26
liblapack=3.9.0=14_osxarm64_openblas
liblapacke=3.9.0=14_osxarm64_openblas
libnghttp2=1.46.0=h95c9599_0
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=hb8d0fd4_0
libssh2=1.10.0=hf27765b_0
libtiff=4.2.0=h11e2f9f_0
libwebp=1.2.2=h68602c7_0
libwebp-base=1.2.2=h1a28f6b_0
llvm-openmp=14.0.3=hd125106_0
lz4-c=1.9.3=hc377ac9_0
markdown=3.3.7=pypi_0
matplotlib=3.5.2=pypi_0
mujoco-py=2.1.2.14=pypi_0
ncurses=6.3=h1a28f6b_2
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
numpy=1.22.3=pypi_0
oauthlib=3.2.0=pypi_0
openblas=0.3.20=openmp_h745f6c2_0
openssl=1.1.1o=h1a28f6b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pickle-mixin=1.0.2=pypi_0
pillow=9.1.1=pypi_0
pip=22.1=pypi_0
protobuf=3.20.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
python=3.8.13=hbdb9e5c_0
python-dateutil=2.8.2=pypi_0
readline=8.1.2=h1a28f6b_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scipy=1.8.1=pypi_0
setuptools=61.2.0=py38hca03da5_0
six=1.15.0=pypi_0
sqlite=3.38.3=h1058600_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.8.0=0
tensorflow-macos=2.8.0=pypi_0
tensorflow-metal=0.4.0=pypi_0
termcolor=1.1.0=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
tk=8.6.11=hb8d0fd4_1
torch=1.11.0=pypi_0
torchvision=0.2.2=py_3
tqdm=4.64.0=pypi_0
typing-extensions=4.2.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.5=h1a28f6b_1
zipp=3.8.0=pypi_0
zlib=1.2.12=h5a0b063_2
zstd=1.4.9=h8574219_1
